pyproject.toml
[tool.poetry]
name = "house-scheduler"
version = "0.1.0"
description = ""
authors = ["Dave Brown <dave.brown@exit107.com>"]
readme = "README.md"
packages = [{include = "house_scheduler"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
cronex = "^0.1.3.1"
pydantic = "^1.10.8"
pytest = "^7.3.1"
croniter = "^1.3.14"
pytz = "^2023.3"
types-pyyaml = "^6.0.12.10"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ipython = "^8.13.2"
pytest = "^7.3.1"
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.2.0"
types-croniter = "^1.3.2.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
